Improve stats collection and throughput information
Make the humanize function accept a float64 rather than converting internally
Collect throughput statistics from writers and total them for each target
Add a verbose flag for output between debug and nearly nothing
Add an option to record per-second block device stats (only Linux)
Add CSV output for block device statistics

+++ buffer_test.go
@@ -75,6 +75,6 @@ func BenchmarkDataReader_Read(b *testing.B) {
 	b.Logf(
 		"Copied %d, %s/sec",
 		totalCopied,
-		sizeHumanizer(totalCopied/int(time.Now().Sub(startTime).Seconds()), true),
+		sizeHumanizer(float64(totalCopied)/time.Now().Sub(startTime).Seconds(), true),
 	)
 }
